backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.1
pydantic-settings==2.8.1
supabase==2.13.0
httpx==0.28.1
litellm==1.63.11
numpy==2.2.4
scipy==1.15.2
pdfplumber==0.11.6
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.1.0
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
Jinja2==3.1.6
aiofiles==24.1.0
